+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## 7.5.11 (Unreleased)
+## 7.6.0 (Unreleased)
 
 ### Bugfixes
 
  - Fix inability of Input custom_options value to be 0
  - Fix improper unit conversion for TH1X-AM2301 Input ([#670](https://github.com/kizniche/mycodo/issues/670))
+ - Fix Bash Command Input script execution ([#667](https://github.com/kizniche/mycodo/issues/667))
 
 ### Features